requirements.txt
Flask
pdfminer.six
python-docx
nltk
scikit-learn
numpy
pandas